Refactored test from #3989 to work correctly and also apply to #4191 and fixed the bug being tested

+++ src_c/draw.c
@@ -2327,9 +2327,9 @@ draw_fillpoly(SDL_Surface *surf, int *point_x, int *point_y,
      * num_points : the number of points
      */
     Py_ssize_t i, i_previous;  // i_previous is the index of the point before i
-    int y, miny, maxy;
-    int x1, y1;
-    int x2, y2;
+    double y, miny, maxy;
+    double x1, y1;
+    double x2, y2;
     /* x_intersect are the x-coordinates of intersections of the polygon
      * with some horizontal line */
     int *x_intersect = PyMem_New(int, num_points);
@@ -2356,7 +2356,8 @@ draw_fillpoly(SDL_Surface *surf, int *point_x, int *point_y,
             minx = MIN(minx, point_x[i]);
             maxx = MAX(maxx, point_x[i]);
         }
-        drawhorzlineclipbounding(surf, color, minx, miny, maxx, drawn_area);
+        drawhorzlineclipbounding(surf, color, minx, (int)miny, maxx,
+                                 drawn_area);
         PyMem_Free(x_intersect);
         return;
     }
@@ -2396,13 +2397,13 @@ draw_fillpoly(SDL_Surface *surf, int *point_x, int *point_y,
                 // add intersection if y crosses the edge (excluding the lower
                 // end), or when we are on the lowest line (maxy)
                 x_intersect[n_intersections++] =
-                    (y - y1) * (x2 - x1) / (y2 - y1) + x1;
+                    (int)((y - y1) * (x2 - x1) / (y2 - y1) + x1);
             }
         }
         qsort(x_intersect, n_intersections, sizeof(int), compare_int);
 
         for (i = 0; (i < n_intersections); i += 2) {
-            drawhorzlineclipbounding(surf, color, x_intersect[i], y,
+            drawhorzlineclipbounding(surf, color, x_intersect[i], (int)y,
                                      x_intersect[i + 1], drawn_area);
         }
     }
@@ -2421,7 +2422,7 @@ draw_fillpoly(SDL_Surface *surf, int *point_x, int *point_y,
         y = point_y[i];
 
         if ((miny < y) && (point_y[i_previous] == y) && (y < maxy)) {
-            drawhorzlineclipbounding(surf, color, point_x[i], y,
+            drawhorzlineclipbounding(surf, color, point_x[i], (int)y,
                                      point_x[i_previous], drawn_area);
         }
     }
